central/requirements.txt
flask==3.0.2
flask-socketio==5.3.6
eventlet==0.35.2
requests==2.31.0
pydantic==2.6.4
pyyaml==6.0.1
